Push down EqualTo through CaseWhen/If
