+++ sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/optimizer/SimplifyConditionalSuite.scala
@@ -199,90 +199,4 @@ class SimplifyConditionalSuite extends PlanTest with ExpressionEvalHelper with P
         If(Factorial(5) > 100L, b, nullLiteral).eval(EmptyRow))
     }
   }
-
-  test("SPARK-33798: Push down EqualTo through If") {
-    val a = EqualTo(UnresolvedAttribute("a"), Literal(100))
-    val b = UnresolvedAttribute("b")
-    val ifExp = If(a, Literal(2), Literal(3))
-
-    assertEquivalent(EqualTo(ifExp, Literal(4)), FalseLiteral)
-    assertEquivalent(EqualTo(ifExp, Literal(3)), If(a, FalseLiteral, TrueLiteral))
-    assertEquivalent(EqualTo(ifExp, Literal("4")), FalseLiteral)
-    assertEquivalent(EqualTo(ifExp, Literal("3")), If(a, FalseLiteral, TrueLiteral))
-
-    // Do not simplify if it contains non foldable expressions.
-    assertEquivalent(
-      EqualTo(If(a, b, Literal(2)), Literal(2)),
-      EqualTo(If(a, b, Literal(2)), Literal(2)))
-
-    // Do not simplify if it contains non-deterministic expressions.
-    val nonDeterministic = If(LessThan(Rand(1), Literal(0.5)), Literal(1), Literal(1))
-    assert(!nonDeterministic.deterministic)
-    assertEquivalent(EqualTo(nonDeterministic, Literal(-1)), EqualTo(nonDeterministic, Literal(-1)))
-
-    // Handle Null values.
-    assertEquivalent(
-      EqualTo(If(a, Literal(null, IntegerType), Literal(1)), Literal(1)),
-      If(a, Literal(null, BooleanType), TrueLiteral))
-    assertEquivalent(
-      EqualTo(If(a, Literal(null, IntegerType), Literal(1)), Literal(2)),
-      If(a, Literal(null, BooleanType), FalseLiteral))
-    assertEquivalent(
-      EqualTo(If(a, Literal(1), Literal(2)), Literal(null, IntegerType)),
-      Literal(null, BooleanType))
-    assertEquivalent(
-      EqualTo(If(a, Literal(null, IntegerType), Literal(null, IntegerType)), Literal(1)),
-      Literal(null, BooleanType))
-  }
-
-  test("SPARK-33798: Push down EqualTo through CaseWhen") {
-    val a = EqualTo(UnresolvedAttribute("a"), Literal(100))
-    val b = UnresolvedAttribute("b")
-    val c = EqualTo(UnresolvedAttribute("c"), Literal(true))
-    val caseWhen = CaseWhen(Seq((a, Literal(1)), (c, Literal(2))), Some(Literal(3)))
-
-    assertEquivalent(EqualTo(caseWhen, Literal(4)), FalseLiteral)
-    assertEquivalent(EqualTo(caseWhen, Literal(3)),
-      CaseWhen(Seq((a, FalseLiteral), (c, FalseLiteral)), Some(TrueLiteral)))
-    assertEquivalent(EqualTo(caseWhen, Literal("4")), FalseLiteral)
-    assertEquivalent(EqualTo(caseWhen, Literal("3")),
-      CaseWhen(Seq((a, FalseLiteral), (c, FalseLiteral)), Some(TrueLiteral)))
-    assertEquivalent(
-      EqualTo(CaseWhen(Seq((a, Literal("1")), (c, Literal("2"))), None), Literal("4")),
-      CaseWhen(Seq((a, FalseLiteral), (c, FalseLiteral)), None))
-
-    assertEquivalent(
-      And(EqualTo(caseWhen, Literal(5)), EqualTo(caseWhen, Literal(6))),
-      FalseLiteral)
-
-    // Do not simplify if it contains non foldable expressions.
-    assertEquivalent(EqualTo(caseWhen, NonFoldableLiteral(true)),
-      EqualTo(caseWhen, NonFoldableLiteral(true)))
-    val nonFoldable = CaseWhen(Seq(normalBranch, (a, b)), None)
-    assertEquivalent(EqualTo(nonFoldable, Literal(1)), EqualTo(nonFoldable, Literal(1)))
-
-    // Do not simplify if it contains non-deterministic expressions.
-    val nonDeterministic = CaseWhen(Seq((LessThan(Rand(1), Literal(0.5)), Literal(1))), Some(b))
-    assert(!nonDeterministic.deterministic)
-    assertEquivalent(EqualTo(nonDeterministic, Literal(-1)), EqualTo(nonDeterministic, Literal(-1)))
-
-    // Handle Null values.
-    assertEquivalent(
-      EqualTo(CaseWhen(Seq((a, Literal(null, IntegerType))), Some(Literal(1))), Literal(2)),
-      CaseWhen(Seq((a, Literal(null, BooleanType))), Some(FalseLiteral)))
-    assertEquivalent(
-      EqualTo(CaseWhen(Seq((a, Literal(1))), Some(Literal(2))), Literal(null, IntegerType)),
-      Literal(null, BooleanType))
-    assertEquivalent(
-      EqualTo(CaseWhen(Seq((a, Literal(null, IntegerType))), Some(Literal(1))), Literal(1)),
-      CaseWhen(Seq((a, Literal(null, BooleanType))), Some(TrueLiteral)))
-    assertEquivalent(
-      EqualTo(CaseWhen(Seq((a, Literal(null, IntegerType))), Some(Literal(null, IntegerType))),
-        Literal(1)),
-      Literal(null, BooleanType))
-    assertEquivalent(
-      EqualTo(CaseWhen(Seq((a, Literal(null, IntegerType))), Some(Literal(null, IntegerType))),
-        Literal(null, IntegerType)),
-      Literal(null, BooleanType))
-  }
 }
